Refactor spark module into transforms and commands
