+++ tests/test_spark.py
@@ -8,3 +8,19 @@ def spark():
     spark.conf.set("spark.sql.session.timeZone", "UTC")
     yield spark
     spark.stop()
+
+
+def test_transform_verify1():
+    raise NotImplementedError
+
+
+def test_transform_verify2():
+    raise NotImplementedError
+
+
+def test_transform_aggregate():
+    raise NotImplementedError
+
+
+def test_transform_publish():
+    raise NotImplementedError
